Added history request.

+++ app/src/main/java/com/smapl_android/net/responses/TrackingResponse.java
@@ -6,19 +6,19 @@ import com.google.gson.annotations.SerializedName;
 public class TrackingResponse extends ServerResponse<TrackingResponse> {
 
     @SerializedName("total_amount")
-    private int totalAmount;
+    private double totalAmount;
 
     @SerializedName("total_distance")
-    private int totalDistance;
+    private double totalDistance;
 
     @SerializedName("status")
     private String status;
 
-    public int getTotalAmount() {
+    public double getTotalAmount() {
         return getResult().totalAmount;
     }
 
-    public int getTotalDistance() {
+    public double getTotalDistance() {
         return getResult().totalDistance;
     }
